+++ Binance.ApiConsole/Program.cs
@@ -317,6 +317,7 @@ internal class Program
         var futures_224 = await api.UsdFutures.GetMarginChangeHistoryAsync("---SYMBOL---");
         var futures_225 = await api.UsdFutures.GetAlgoOrderAsync(algoId: 1_000_001L);
         var futures_226 = await api.UsdFutures.GetOpenAlgoOrdersAsync(symbol: "---SYMBOL---");
+        var futures_227 = await api.UsdFutures.GetAlgoOrdersAsync("---SYMBOL---");
 
         // USDⓈ-M Futures -> User Data Stream Methods (PRIVATE)
         var futures_301 = await api.UsdFutures.StartUserStreamAsync();